Inject GPU into region.capabilities for Dallas
