+++ packages/manager/src/store/regions/regions.actions.ts
@@ -1,4 +1,4 @@
-import { getRegions, Region } from '@linode/api-v4/lib/regions';
+import { Capabilities, getRegions, Region } from '@linode/api-v4/lib/regions';
 import { APIError } from '@linode/api-v4/lib/types';
 import { actionCreatorFactory } from 'typescript-fsa';
 
@@ -17,7 +17,9 @@ export const requestRegions: RequestRegionsThunk = () => dispatch => {
   dispatch(regionsRequestActions.started());
   return getRegions()
     .then(regions => {
-      dispatch(regionsRequestActions.done({ result: regions.data }));
+      dispatch(
+        regionsRequestActions.done({ result: regions.data.map(addGPUToRegion) })
+      );
       return regions;
     })
     .catch(error => {
@@ -25,3 +27,16 @@ export const requestRegions: RequestRegionsThunk = () => dispatch => {
       return error;
     });
 };
+
+/**
+ * One day soon, the API will return GPU as one of a region's capabilities.
+ * Until that, we're faking it here.
+ */
+const regionsWithGPUs = ['us-central'];
+const addGPUToRegion = (region: Region) =>
+  regionsWithGPUs.includes(region.id)
+    ? {
+        ...region,
+        capabilities: [...region.capabilities, 'GPU'] as Capabilities[]
+      }
+    : region;
